Add issuer email address field to certificates
We're using a multiselect field with the 'creatable' and 'searchable'
attributes in order to allow users to type in an arbitrary email address.

In the back-end, these email addresses are used to look up users and
create appropriate records in the certificate-issuer join table,
and always the author themselves is included.

Any unrecognised email addresses are silently dropped.